ci(deployment): partially done
