--- ormconfig.js
@@ -1,30 +1,0 @@
-var dbConfig = {
-  synchronize: true,
-  migrations: ['migrations/*.js'],
-  cli: {
-    migrationsDir: 'migrations',
-  },
-};
-
-switch (process.env.NODE_ENV) {
-  case 'development':
-    Object.assign(dbConfig, {
-      type: 'sqlite',
-      database: 'db.sqlite',
-      entities: ['**/*.entity.js'],
-    });
-    break;
-  case 'test':
-    Object.assign(dbConfig, {
-      type: 'sqlite',
-      database: 'test.sqlite',
-      entities: ['**/*.entity.ts'],
-    });
-    break;
-  case 'production':
-    break;
-  default:
-    throw new Error('Unknown environment');
-}
-
-module.exports = dbConfig;
