+++ package.json
@@ -18,7 +18,11 @@
     "test:cov": "cross-env NODE_ENV=test jest --coverage",
     "test:debug": "cross-env NODE_ENV=test node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "cross-env NODE_ENV=test jest --config ./test/jest-e2e.json",
-    "typeorm": "cross-env NODE_ENV=development node --require ts-node/register ./node_modules/typeorm/cli.js"
+    "typeorm": "cross-env NODE_ENV=development ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js -d src/scripts/db/data-source-developement.ts",
+    "migration:generate": "npm run typeorm -- migration:generate migrations/Migration",
+    "migration:migrate": "npm run typeorm -- migration:run",
+    "migration:revert": "npm run typeorm -- migration:revert",
+    "das": "npm run build && typeorm migration:generate src/migrations/Migration npx typeorm -d dist/scripts/db/data-source-development.js "
   },
   "dependencies": {
     "@nestjs/common": "^10.0.0",
