+++ scripts/db/data-source-test.ts
@@ -1,0 +1,9 @@
+import { DataSourceOptions } from 'typeorm';
+
+export const dbConfig: DataSourceOptions = {
+  type: 'sqlite',
+  database: 'test.sqlite',
+  entities: ['**/*.entity.js'],
+  logging: false,
+  migrations: ['dist/datasource/migrations/*.js'],
+};
